move unbond config to common
