+++ guild-sc-config/Cargo.toml
@@ -11,6 +11,10 @@ path = "src/lib.rs"
 [dependencies.multiversx-sc]
 version = "=0.47.5"
 
+[dependencies.common_structs]
+git = "https://github.com/multiversx/mx-exchange-sc"
+rev = "71d5305"
+
 [dev-dependencies]
 num-bigint = "0.4"
