+++ guild-sc/src/user_actions/unstake_farm.rs
@@ -133,7 +133,7 @@ pub trait UnstakeFarmModule:
         amount: BigUint,
         original_attributes: StakingFarmTokenAttributes<Self::Api>,
     ) -> EsdtTokenPayment {
-        let min_unbond_epochs = self.min_unbond_epochs().get();
+        let min_unbond_epochs = self.get_min_unbond_epochs();
         let current_epoch = self.blockchain().get_block_epoch();
 
         self.unbond_token().nft_create_and_send(
