+++ guild-sc/tests/farm_staking_setup/mod.rs
@@ -96,7 +96,7 @@ where
 
         b_mock
             .execute_tx(&owner_addr, &config_wrapper, &rust_zero, |sc| {
-                sc.init(managed_biguint!(i64::MAX));
+                sc.init(managed_biguint!(i64::MAX), MIN_UNBOND_EPOCHS);
 
                 let mut tiers = MultiValueEncoded::new();
                 tiers.push(
@@ -123,7 +123,6 @@ where
                 sc.init(
                     farming_token_id,
                     division_safety_constant,
-                    MIN_UNBOND_EPOCHS,
                     managed_address!(&owner_addr),
                     managed_address!(config_wrapper.address_ref()),
                     managed_address!(&owner_addr),
